pyproject.toml
[project]
name = "trust-metric-aggregator"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prometheus-client (>=0.21.1,<0.22.0)",
    "strawberry-graphql[fastapi] (>=0.265.1,<0.266.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
